move metamask and rm un-needed package
