+++ packages/commonwealth/package.json
@@ -176,7 +176,6 @@
     "css-minimizer-webpack-plugin": "^4.2.2",
     "dompurify": "^2.2.6",
     "dotenv": "^8.2.0",
-    "eth-sig-util": "^3.0.1",
     "ethereum-block-by-date": "^1.2.2",
     "ethereumjs-util": "7.1.0",
     "ethereumjs-wallet": "^1.0.2",
