+++ packages/commonwealth/test/unit/wallets/metamask.spec.ts
@@ -1,6 +1,6 @@
-import MetamaskWebWalletController from "client/scripts/controllers/app/webWallets/metamask_web_wallet";
+import MetamaskWebWalletController from "../../../client/scripts/controllers/app/webWallets/metamask_web_wallet";
 import { ethers } from "ethers";
-import { getMetaMaskMock } from "test/e2eStateful/mockMetaMaskUtil";
+import { getMetaMaskMock } from "../../util/mockMetaMaskUtil";
 
 const window = {ethereum: getMetaMaskMock('https://mainnet.infura.io/v3/17253b2fd784479abff55a32c9b3282c', ethers.Wallet.createRandom().privateKey)};
